Add files via upload
Term deposits are a major source of income for a bank.
A term deposit is a cash investment held at a financial institution. Your
money is invested for an agreed rate of interest over a fixed amount of
time, or term.
The bank has various outreach plans to sell term deposits to their
customers such as email marketing, advertisements, telephonic marketing
and digital marketing.
Telephonic marketing campaigns still remain one of the most effective way
to reach out to people. However, they require huge investment as large call
centers are hired to actually execute these campaigns. Hence, it is crucial
to identify the customers most likely to convert beforehand so that they can
be specifically targeted via call.
 The client data such as : age of the client, their job
type, their marital status, etc. Along with the client data,  the information of the call such as the duration of the call, day and month of the call, etc. are given , task is to predict if
the client will subscribe to term deposit.